Converts `txt_embedding` tensor to double

+++ kilbert/vilbert/vilbert.py
@@ -502,8 +502,8 @@ class CustomBertSelfAttention(nn.Module):
         
     def forward(self, hidden_states, attention_mask, txt_embedding_kb):
         mixed_query_layer = self.query(hidden_states)
-        mixed_key_layer = self.key(txt_embedding_kb)
-        mixed_value_layer = self.value(txt_embedding_kb)
+        mixed_key_layer = self.key(txt_embedding_kb.double())
+        mixed_value_layer = self.value(txt_embedding_kb.double())
         
         query_layer = self.transpose_for_scores(mixed_query_layer)
         key_layer = self.transpose_for_scores(mixed_key_layer)
